Updated receiver
replaced array-list with hash map, as array-list was being to difficult with iterations, and the map was much easier to deal with. 
first block of code after listen block is checking new segment against map, and adding into the hash map if it is a new value. 
second block of code is frankly, optional. it checks if it is able to match value of ackNo with received segments seq. value, and adding length value to AckNo. 

will need to adjust else { while{ loop location, as it will start that process too early if i receive a segment thats already in the map, or need to add a counter to ensure all files are received. 

while loop counts down segments received timer and immediately finds the matching ackNo to sequence number, and adds that length value and repeats, until TotalSegments is 0 ( which ideally will be at 10, after hopefully getting all of the segments).